+++ test/integration/http_integration.h
@@ -159,6 +159,11 @@ protected:
   void testDrainClose();
   void testRetry();
   void testRetryHittingBufferLimit();
+  IntegrationStreamDecoderPtr setupPerStreamIdleTimeoutTest();
+  void testPerStreamIdleTimeoutBeforeUpstreamHeaders();
+  void testPerStreamIdleTimeoutAfterUpstreamHeaders();
+  void testPerStreamIdleTimeoutAfterData();
+  void testPerStreamIdleTimeoutRequestAndResponse();
   void testGrpcRouterNotFound();
   void testGrpcRetry();
   void testHittingDecoderFilterLimit();